// src/utilities/Types.ts
export type lookup<T> = {
  [key: string]: T;
};

export type ShotchartZoneCSS =
  | "right-corner-three-zone"
  | "left-corner-three-zone"
  | "right-three-zone"
  | "left-three-zone"
  | "middle-three-zone"
  | "right-baseline-midrange-zone"
  | "left-baseline-midrange-zone"
  | "right-wing-midrange-zone"
  | "left-wing-midrange-zone"
  | "middle-midrange-zone"
  | "left-floater-zone"
  | "right-floater-zone"
  | "middle-floater-zone"
  | "rim-zone";

export type ShotchartZone =
  | "R-C3"
  | "L-C3"
  | "R-ATB"
  | "L-ATB"
  | "M-ATB"
  | "RB-MR"
  | "LB-MR"
  | "RW-MR"
  | "LW-MR"
  | "M-MR"
  | "L-FL"
  | "R-FL"
  | "M-FL"
  | "RIM";